# src/api_service/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API с JWT авторизацией",
    "description": "Пример API, использующего JWT для авторизации в Swagger UI",
    "version": "1.0"
  },
  "paths": {
...
    "/v1/user/profile": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get Profile",
        "operationId": "get_profile_v1_user_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticate",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Not authenticate"
                }
              }
            }
          },
          "404": {
            "description": "Entity not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Entity not found"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/user/list": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "List Users",
        "operationId": "list_users_v1_user_list_get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDTO"
                  },
                  "title": "Response List Users V1 User List Get"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticate",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Not authenticate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/user/auth": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Auth User",
        "operationId": "auth_user_v1_user_auth_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelegramAuthDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "example": {
                  "token": "REDACTED"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/v1/stats/coworking-count": {
      "get": {
        "tags": [
          "Stats"
        ],
        "summary": "Get Coworking Count",
        "operationId": "get_coworking_count_v1_stats_coworking_count_get",
        "responses": {
          "200": {
            "description": "Returns count of coworkings",
            "content": {
              "application/json": 1
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/stats/workplaces-count": {
      "get": {
        "tags": [
          "Stats"
        ],
        "summary": "Get Workplaces Count",
        "operationId": "get_workplaces_count_v1_stats_workplaces_count_get",
        "responses": {
          "200": {
            "description": "Returns count of workplaces",
            "content": {
              "application/json": 9
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/stats/medium-price-per-hour": {
      "get": {
        "tags": [
          "Stats"
        ],
        "summary": "Get Medium Price Per Hour",
        "operationId": "get_medium_price_per_hour_v1_stats_medium_price_per_hour_get",
        "responses": {
          "200": {
            "description": "Returns medium price per hour",
            "content": {
              "application/json": 7461.42
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/stats/occupancy-rate": {
      "get": {
        "tags": [
          "Stats"
        ],
        "summary": "Get Occupancy Rate",
        "operationId": "get_occupancy_rate_v1_stats_occupancy_rate_get",
        "responses": {
          "200": {
            "description": "Returns occupancy rate",
            "content": {
              "application/json": 7.5
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/v1/external/tg_admin/add": {
      "post": {
        "tags": [
          "External"
        ],
        "summary": "Add Tg Admin",
        "operationId": "add_tg_admin_v1_external_tg_admin_add_post",
        "parameters": [
          {
            "name": "tg_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tg Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of tg admins",
            "content": {
              "application/json": {
                "schema": {},
                "example": [
                  1282629807,
                  1522105862,
                  5367427116,
                  946082604,
                  256086824
                ]
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/external/tg_admin/remove": {
      "delete": {
        "tags": [
          "External"
        ],
        "summary": "Remove Tg Admin",
        "operationId": "remove_tg_admin_v1_external_tg_admin_remove_delete",
        "parameters": [
          {
            "name": "tg_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tg Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of tg admins",
            "content": {
              "application/json": {
                "schema": {},
                "example": [
                  1282629807,
                  1522105862,
                  5367427116,
                  946082604
                ]
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TelegramAuthDTO": {
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 1,
            "title": "Id",
            "examples": [
              1522105862
            ]
          },
          "first_name": {
            "type": "string",
            "minLength": 1,
            "title": "First Name",
            "examples": [
              "Дмитрий"
            ]
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0,
                "examples": [
                  "Нагиев"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "username": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "examples": [
                  "parapopovich"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0,
                "examples": [
                  "https://t.me/i/userpic/320/7IQwLdFc8nusS8VT2ll1tQnEW8pKzd3wP-s_SbiBsmY.svg"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "auth_date": {
            "type": "integer",
            "minimum": 0,
            "title": "Auth Date",
            "examples": [
              1742720967
            ]
          },
          "hash": {
            "type": "string",
            "minLength": 1,
            "title": "Hash",
            "examples": [
              "035da2b6954c0e8dfd8de9a10b6644152a66c03aadda72205fec89806e0fc9b3"
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "first_name",
          "auth_date",
          "hash"
        ],
        "title": "TelegramAuthDTO"
      },
      "UserDTO": {
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 1,
            "title": "Id",
            "examples": [
              1522105862
            ]
          },
          "first_name": {
            "type": "string",
            "minLength": 1,
            "title": "First Name",
            "examples": [
              "Дмитрий"
            ]
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0,
                "examples": [
                  "Нагиев"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "username": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "examples": [
                  "parapopovich"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0,
                "examples": [
                  "https://t.me/i/userpic/320/7IQwLdFc8nusS8VT2ll1tQnEW8pKzd3wP-s_SbiBsmY.svg"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "balance": {
            "type": "integer",
            "minimum": 0,
            "title": "Balance",
            "examples": [
              4252
            ]
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "examples": [
              "2025-03-04T06:37:03.960093"
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "first_name",
          "balance",
          "role",
          "created_at"
        ],
        "title": "UserDTO"
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "USER",
          "ADMIN"
        ],
        "title": "UserRole"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT токен для авторизации. Формат: Bearer <токен>"
      }
    }
  },
  "servers": [
    {
      "url": "/api"
    }
  ]
}